requirements_enhanced.txt
# Enhanced XelBot Requirements - Professional Data Analytics Chatbot

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# AI & Machine Learning
google-generativeai==0.3.2
openai==1.3.7

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Business Intelligence & Reporting
reportlab==4.0.7
jinja2==3.1.2
openpyxl==3.1.2

# Web & API
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Optional: Database support (if needed later)
# sqlalchemy==2.0.23
# sqlite3 (built-in)

# Optional: Advanced ML (if needed later)
# tensorflow==2.15.0
# torch==2.1.1
